+++ app/src/main/java/com/flowfoundation/wallet/page/backup/multibackup/viewmodel/BackupRecoveryPhraseViewModel.kt
@@ -8,7 +8,6 @@ import com.flowfoundation.wallet.manager.account.DeviceInfoManager
 import com.flowfoundation.wallet.manager.backup.BackupCryptoProvider
 import com.flowfoundation.wallet.manager.flowjvm.CadenceScript
 import com.flowfoundation.wallet.manager.flowjvm.transactionByMainWallet
-import com.flowfoundation.wallet.manager.flowjvm.ufix64Safe
 import com.flowfoundation.wallet.manager.transaction.OnTransactionStateChange
 import com.flowfoundation.wallet.manager.transaction.TransactionState
 import com.flowfoundation.wallet.manager.transaction.TransactionStateManager
@@ -43,6 +42,7 @@ import com.flowfoundation.wallet.utils.Env.getStorage
 import org.onflow.flow.ChainId
 import com.flowfoundation.wallet.manager.account.AccountWalletManager
 import com.flowfoundation.wallet.manager.wallet.WalletManager
+import org.onflow.flow.infrastructure.Cadence.Companion.uint8
 
 class BackupRecoveryPhraseViewModel : ViewModel(), OnTransactionStateChange {
     val createBackupCallbackLiveData = MutableLiveData<Boolean>()
@@ -109,8 +109,8 @@ class BackupRecoveryPhraseViewModel : ViewModel(), OnTransactionStateChange {
                 try {
                     val txId = CadenceScript.CADENCE_ADD_PUBLIC_KEY.transactionByMainWallet {
                         arg { string(it.getPublicKey()) }
-                        arg { uint8(it.getSignatureAlgorithm().cadenceIndex) }
-                        arg { uint8(it.getHashAlgorithm().cadenceIndex) }
+                        arg { uint8(it.getSignatureAlgorithm().cadenceIndex.toUByte()) }
+                        arg { uint8(it.getHashAlgorithm().cadenceIndex.toUByte()) }
                         arg { ufix64Safe(500) }
                     }
                     val transactionState = TransactionState(
