+++ app/src/main/java/com/flowfoundation/wallet/page/staking/amount/dialog/StakingAmountConfirmDialog.kt
@@ -25,6 +25,7 @@ import com.flowfoundation.wallet.utils.extensions.res2String
 import com.flowfoundation.wallet.utils.extensions.setVisible
 import com.flowfoundation.wallet.widgets.ButtonState
 import kotlinx.coroutines.delay
+import org.onflow.flow.infrastructure.Cadence.Companion.uint32
 
 class StakingAmountConfirmDialog : BottomSheetDialogFragment() {
 
@@ -110,7 +111,7 @@ class StakingAmountConfirmDialog : BottomSheetDialogFragment() {
             }
             val txId = CadenceScript.CADENCE_STAKE_FLOW.transactionByMainWallet {
                 arg { string(data.provider.id) }
-                arg { uint32(delegatorId) }
+                arg { uint32(delegatorId.toUInt()) }
                 arg { ufix64Safe(amount) }
             }
             val transactionState = TransactionState(
@@ -143,7 +144,7 @@ class StakingAmountConfirmDialog : BottomSheetDialogFragment() {
             }
             val txId = CadenceScript.CADENCE_UNSTAKE_FLOW.transactionByMainWallet {
                 arg { string(data.provider.id) }
-                arg { uint32(delegatorId) }
+                arg { uint32(delegatorId.toUInt()) }
                 arg { ufix64Safe(amount) }
             }
             val transactionState = TransactionState(
